Prototype - Added basic structure
Prototype - Added basic structure for indicators placement and the manner of execution.
By this way, for the setting i am currently using, there are 25 configurable variables. That means the execution will be of the order O(n^25). This will consume a lot of time, may be more than my lifetime. We will optimize the input length so that execution time can be controlled.
Other wise we will have to follow something like waterfall model - find the best / optimal settings for 1 indicator and improve the optimality by building over the existing solution.